fix: low level rejection of too many colliders
